Move formatting logic to separate class on Android
